Restrict ALL_VOLUMES query parameter to <= R.
This parameter was added for supporting existing logic by partners
to keep their own copy of a media database in sync. For now, limit this
to R and lower only, until we know if (and what kind of) API we want for
this.

Bug: 171025568
Test: N/A
Change-Id: I8af4407cc3284877de3135a267a5524ac72655ab
(cherry picked from commit c7b11aa2cf951140c3b1b4e8532bc6d00cac7523)