Modify test to remove flaky assertions
Insert in a playlist returns the member uri appended with
row id of audio_playlists_map, and not the play order id or audio_id.
Hence, we should not be making this assertion as it could vary on
the number of rows in the db.

Bug: 173703755
Test: atest ClientPlaylistTest
Change-Id: Ie732fd69f6bd420ebe114341b154912f11842f8f
Merged-In: Ie732fd69f6bd420ebe114341b154912f11842f8f
(cherry picked from commit feb6c274fa7f65ad3df0920aae3b0ae9f91f1872)