Modify PerformanceTests to test for renames and greater size
* Add 2 rename timers to the set of directory operations tests that we
already have. 2 timers are essential as we want to track performance for
renaming a directory and files separately.
* Bring back test directory operations for size: 500. We could not add
these before as it would take a lot of time ag/11951722.
(It takes approx 20s after the db bypass CL which was merged
later).
* Add LargeTest to track performance for dir size 1000. This test will
not run in presubmit, as it gets exceeds the 60s timeout.
Add this test to run in postsubmit as the time duration increases
exponentially (it takes approx 3-5 times the duration for size 500) and
it would be helpful to track the performance for this.

Bug: 173156920
Test: atest com.android.providers.media.client.PerformanceTest
Test: atest --test-mapping packages/providers/MediaProvider:presubmit
Test: atest --test-mapping packages/providers/MediaProvider:postsubmit

Change-Id: I66a738f25e24dc741d79cab8445301eaef6e632d