Fix path used for redaction when transcoding
The node->BuildPath() is the path accessed by app and should be used
for db operations.
The node->GetIoPath() is the path required for IO (might be a
transcoded path not indexed by the db)

Previously, we passed only the io_path to Java, so for transcoding, we
would run into errors since MediaProvider can't find that path in the
db.

Now, we pass both paths to Java and use io_path for IO and
build_path (|path|) for db ops

Bug: 171953356
Test: atest TranscodeTest (should be a no-op without transcoding)
Change-Id: I88f63b0490752717ef8ea65d8df4d117b4f95ea0