Fix transcoding status update for ContentResolver requests
When a transcoding job completes successfully, we update a database
row for transcode_status to 'completed'. This update was failing
because extra SQL 'WHERE' clauses were attached to the update SQL to
match the calling apps owner_package_name.

This was a problem only for ContentResolver updates because in those
cases we were running as the calling identity of the caller.

-Now we force all transcode status updates to run as the MediaProvider
calling identity.

-Added @Ignore regression test currently blocked by b/174655855

-Rewrote the assertTranscode util to only check read(2) duration. This
should be more reliable across file path and ContentResolver
especially as open latency can be very different in either case.

Bug: 174521420
Test: atest TranscodeTest
Change-Id: Ibb71876b5c92a730442398bff4ba950a96b5b3f4