Fix checking device_config in MediaProvider
There are 2 problems:
1. Running as real MediaProvider: When app calls ContentResolver#open,
we are running as the Binder calling identity of the app. In order to check
transcode device configs, we need to clear the calling identity before
calling into system_server
2. Running as instrumented MediaProvider: During some tests we create
an IsolatedContext and instantiate a MediaProvider instance running as
the test uid. Since the test doesn't have device_config permissions,
we mock the device_config check to return a default value and not
actually call into system_server

Bug: 169327180
Test: presubmit
Change-Id: I3500aed035e9783dc1bf939f123fa6cce704ffa9