Fix getVolumeName with secondary volume issue in MediaStore
Since this method doesn't have a context. we can't find the actual
StorageVolume for the given path, so only a vague guess is returned.
If the path is the path of secondary volume, try to return the uuid
from the path.

Test: atest MediaStore_Audio_MediaTest#testGetContentUriForPath
Test: atest --test-mapping packages/providers/MediaProvider
Bug: 174288420
Change-Id: I129c49d4fb8d68ad34444abe2dc5c435747a8161
Merged-In: I129c49d4fb8d68ad34444abe2dc5c435747a8161