Ensure m4a files get the right metadata on scan
We allow changes from video/mp4 to audio/mp4 mime types, but used to do
this when scanning the details of the file. Now, allow this change as
early as possible, to ensure the file is scanned as the correct type
(audio) and gets the appropriate metadata. (e.g. is_music).

Bug: 171983792
Bug: 176522651
Test: atest ModernMediaScannerTest#testScan_148316354
Change-Id: Ia68a425e7f6d15219ccc6561577b39eae77d470d