Improve FUSE file open JNI check
Currently, during FUSE opens, we make a JNI call into the
MediaProvider to check if the uid is allowed to open the specified
path.

This change refactors that check to return a richer FileOpenResult,
the goal is to return the 'original' uid requesting access (perhaps
via MediaProvider APIs). For now we just return the same uid,
status (previous return value) and the redaction ranges (saving an
extra getRedactionRanges call later during open

Bug: 174655855
Test: presubmit

Change-Id: Id86d7f77a736fc201984e132f170451e1aa60697