Add fields to FileLookupResult and improve node invalidation logic
1. transforms_supported field replaces the FLAG_TRANSFORM_SUPPORTED
flag that was added to the transforms field to indicate that
transforms are supported on an inode but the node itself might not
need any transforms. This allows the FUSE daemon to avoid caching that
node for *any* uid. Extracting this flag is a cleaner design.
2. uid field represents the original uid requesting IO. This is useful
when the MediaProvider does a FUSE open on behalf of another app and
we need to know the original uid in the FUSE daemon. This change just
adds the field, the logic for the original uid will be added in later
cl.

3. Refactored the logic to decide if a node should be invalidated in
the VFS dentry cache. Now the node itself contains that information as
a combination of case-insensitivity and transforms_supported signals

Bug: 174655855
Test: presubmit
Change-Id: I3a1b3f5e239289f8d972642e00ec66e6908894bd