Use the media capabilities uid, if specified
TranscodeHelper should check the uid of the app for which this media is
intended for, instead the uid of the app which is opening the media on
behalf of the intended app, to make the decisions to transcode or not
transcode.
When coming via ContentResolver, the calling identity is cleared and we
end up in MediaProvider with its own identity. In such cases we want to
be checking the indentity of the calling app.

BUG: 176157500, 154734139
Test: unit tests added. atest MediaProviderTranscodeTests

Change-Id: I2b5bdd5533dbc0aa39012622b63937d559c05fff