Use manifest attribute for shouldBypass decision
To improve the performance of bulk file path operations like
creating/deleting 1000 files, we allowed apps which are common to have
such use cases such as apps which hold MANAGE_EXTERNAL_STORAGE
permission or apps with System Gallery role to bypass database
operations.

We added new APIs which provide option for file manager apps and system
gallery apps to specify if they would like to skip bypassing database.
The default value of flag is targetSDK gated, and preset while
parsing the app manifest.

Added logic to read the manifest attribute and save it in
LocalCallingIdentity.

Bug: 178209446
Test: atest  android.scopedstorage.cts.device.BypassDatabaseOperationsTest
Change-Id: Ia7c69158b5fa46fee8580e34bc23376793cd1b01