Introduce StatsdPuller
Right now just providing a basic boilerplate for handling the pulling of
metrics by statsd.

To use:
Register STATS_PULL_CALLBACK_HANDLER as the callback handler for your
tag.
Handle the tag appropriately in StatsPullCallbackHandler#onPullAtom().

One way to handle a tag is to have your class for collecting the
metrics, and have static method to handle the responsibility of filling
in the StatsEvent data.

BUG: 160260026, 178813586
Test: builds. This is just a skeleton which doesn't really do anything.

Change-Id: Ia379a5b190d524edba615e2529253bb76510f6eb