Improve transcoding metrics logging
Populated these new fields
1. video_duration
2. capture_framerate
3. transcode_reason

transcode_reason is particularly complicated because we decide the
transcode_reason at FUSE_LOOKUP, but we only log the metric at
FUSE_READ. Hence, a new field was introduced on struct
node#transforms_reason. The transforms_reason is an opaque value like
transforms and it should not be parsed by the native FUSE daemon but
should just be sent along during the transform() call.

Refactored shouldTranscode() to return the transcode reason enum
int value or 0 if should not transcode

Lastly, fixed the metrics logging of read_cache and read_direct.
Previously, we were logging in shouldTranscode which is called very
often in FUSE_LOOKUP. Now, we log during open(2). This is much better,
but is still not technically correct since an app might open but never
read.

Test: atest TranscodeTest
Bug: 178447612

Change-Id: I384f7283e1c4cd6732aca769404c640e412a33f8