Provide infra for pulling of transcode metrics
We are currently using the push mechanism to get our metrics to
westworld. But we wanna move to the pull mechanism.

Here we keep a limit on the total count of data that we store in memory.
And we also keep a limit on sampled data that'd eventually be shared in
the pull callback.

BUG: 160260026, 178813586
Test: atest TranscodeMetricsTest

Change-Id: Ic715b6e7745040c92669f2d8554a37a15a2941de
(cherry picked from commit 34590f2e69fbe72362ea776bc5fea5df07a33396)