Add rename File API support for media URI grants
This change allows apps to rename a file if they have media URI grant
for the file.
This change leaves positive rename case performance unaffected, as the
new logic executes right before returning error.

This change does not support replace. Without this change we currently
support the following (Rename oldFile to newFile):
1. Check if caller has write access for file to be renamed (oldFile).
2. Update oldFile row in db with newFile details.
3. 2 will fail if newFile already exists in the db.
4. Replace scenario kicks in. We check if the caller has write access
for newFile.
5. If 4 is true, we delete db row for newFile and try 2 again.

With this change, we support 1 and check for otherUri grants for
oldFile, but we do not support otherUri grants for replace scenario by
checking if the caller has otherUri grants for newFile. This is done as
the caller should delete the newFile themselves if they have explicit
otherUri grants access for newFile. Created a bug to discuss the
priority of this issue.

Bug: 173648980
Test: atest ExternalStorageHostTest
Change-Id: If143bd4aa29820ce9974ea491dddba51dcc915d7
Merged-In: If143bd4aa29820ce9974ea491dddba51dcc915d7