Skip scanning non-changed audio/mp4 files.
audio/mp4 files can have different extension. If we don't look at the
metadata of the file, we will assume these have changed (because they
have a different mime type) and rescan them. For these files, we now
look at the metadata too. We don't want to always look at the metadata
of files, since it is an expensive operation.

Bug: 176522651
Test: atest MediaProviderTests:ModernMediaScannerTest#testScan_audioMp4_notRescan
Change-Id: Ife87f9e3df83d26c4b614c79f8b96bdb6f3be82c