Do an early return for mediaCapabilitiesUid
for a cleaner looking code.

Also, changed logVerbose to Log.d for logging the case of ignoring
mediaCapabilitiesUid since in most of the cases when a bundle is
supplied we'd expect mediaCapabilitiesUid to be not set (in which case
we'd end up on this log line and wouldn't wanna have this log spuriously
popping up).

BUG: 176157500, 154734139
Test: build
Change-Id: I8fc43b967ebd5f7a39a9308c5478b830f62b9f2e