Fix LocalCallingIdentity callingPackage
In some situations, MediaProvider can get called from the
system_server with a null package name.

Previously, we were assuming MediaProviders package name, but this
could lead to a security exception when we resolve the apps uid and it
doesn't match the package name.

Now, we check if the UID is system and assume we are running as
ourself and not handling any IPC

Bug: 178543548
Bug: 148654029
Test: Manual

Change-Id: I01bcf1aa7645cdaecdc3908216c6eef9a74af5ad