Add QUERY_ARG_DEFER_SCAN systemApi
The flag will indicate if the media scan that was triggered as part of
ContentResolver#update should be asynchronous. This flag should
only be used when ContentResolver#update operation needs to return
early without updating metadata for the file. This may make other
apps see incomplete metadata for the updated file as scan runs
asynchronously here.
When the flag is set, the published file will not appear in default
query until the deferred scan is complete.

Bug: 180326732
Test: presubmit
Change-Id: I1bcdd25c589e25a78f8d2cb501061fd247d550a3