Add test that ORIENTATION column is successfully migrated
While database migration, we migrate a few columns from legacy media
provider as is, but we depend on a scan after migration to update
metadata columns. ORIENTATION is an example of such a column, after
migration to the new db, we wait for a scan to complete to update
ORIENTATION column from file metadata. As scan can take a long time to
complete (depending on the number of files to be scanned), partners
complained that the orientation of a few thumbnails was incorrect
between the migration and the next scan. We fix this by migrating the
ORIENTATION column, which will be eventually replaced on the next scan
by file metadata.

Add poll for cursor as it is flaky sometimes while querying
external_primary. This adds to better code-health.

Bug: 171196888
Test: atest LegacyProviderMigrationTest --iteration 10
Test: This test fails without the respin fix I708686c771b60d1f83f662097fa45a8ad8a2e744
Change-Id: Id167f4451eaa144af310e6db9008f220d615b5ff