Dump finished transcoding sessions
We dump only the most recent
MAX_FINISHED_TRANSCODING_SESSION_STORE_COUNT sessions for each
successful, cancelled, or error'ed sessions. We can change this value in
the future when we realize better.

We keep the successfully finished, cancelled, and error'ed transcoding
sessiond separately, because the probability of getting cancelled or
error'ed sessions is pretty low. And we are keeping a limit on how many
sessions info we store locally. So, we don't wanna miss out on dumping
cancelled or error'ed sessions just because they got cleared from the
local storage.

Keeping a map of the path and the TranscodeSession, instead of creating
a new class and then keeping an array of that class, and then
maintaining a list of most recent entries. Using the map provides an
efficient and clean and simple solution.

BUG: 176882062
Test: presubmit
Change-Id: I126e4dac41673b9d085a01ff6940732f3bf09706