Remove scope for test flakiness for checking ORIENTATION column
Previous test coverage for ORIENTATION column could be flaky after db
migration, as if a scan happens before the assertions for ORIENTATION,
then ORIENTATION column would be set to null. This is because
ORIENTATION column is reset by the metadata value of the file.

Add orientation_90.jpg, and test that the value of ORIENTATION column
should always be 90.

Bug: 171196888
Test: atest com.android.providers.media.client.LegacyProviderMigrationTest#testLegacy_Orientation --iterations 10
Change-Id: Ic8f546607db49d22fd457755d56f333dc240731c