Catch OOME when allocating byte array.
We had reports of OOME during normal usage. We do not have the files to
reproduce, but the error is thrown when allocating a byte array to read
the Iso box. We had a check on MAX_INT, but device can run out of memory
even for smaller values. Added a try-catch to avoid the error.

Bug: 178411011
Test: m; atest .
Change-Id: I60fe57a8c7018dc71cd2ba82cc6d8f68ac39b277