Add implementation for exif redacted Uri APIs
Added a new hidden database column for storing REDACTED_URI_ID,
REDACTED_URI_ID stores the id for redacted uri. Once generated this
field can be reused and hence is only required to be populated once. If
REDACTED_URI_ID is null getRedactedUri will generate an id and store it
in the files DB.

Bug: 179914046
Bug: 176547697

Test: atest CtsScopedStorageDeviceOnlyTest:ScopedStorageDeviceTest

Change-Id: Ic321d239fd0e51dbe3efb302a7f4ac99e8932764