Allow caching FUSE inode attributes
There are several cases where we want to prevent caching dentries in
the kernel so that we can force a FUSE_LOOKUP the next time we try to
access the dentry.

We can set the entry_timeout and attr_timeout to 0 respectively
prevent caching dentries and attributes

Previously, we also prevented caching inode attributes as well in all
the cases where we prevented dentry caching. This was bad and
unneccessary:

-Bad because this caused very frequent GETATTRs when reading from a file
-Unnecessary because the kernel anyways ignored the size and several
other important attributes from the FUSE daemon if write back cache is
enabled.

Now, we just rely on invalidating only the dentry (which consequently
invalidates the attributes) for correctness and performance.

Test: atest ScopedStorageDeviceTest (for case-insensitivity)
Test: atest TranscodeTest
Bug: 171166721
Change-Id: Icad1e4e61e191dca18624ec2e5f3b8310d803837