Support redacted URI file open on FUSE
FUSE needs to be modified so as to be able to open synthetic file
paths. While we want to make sure open on synthetic paths goes fine,
we don't want to expose ".transforms" dir. This CL makes sure that
open on synthetic files goes fine and any sort of modifications to the
".transforms" dir are not allowed.

Test: atest CtsScopedStorageDeviceOnlyTest:ScopedStorageDeviceTest#testRedactionApis -c
Bug: 179914046
Bug: 176547697

Change-Id: Ie4b6d09a0901356cf4492cb8db90caacae3ad13a