Add QUERY_ARG_INCLUDE_RECENTLY_UNMOUNTED_VOLUME API
Added a flag that indicates that ContentResolver#query should include
database rows of recently unmounted volume which are still present in
MediaProvider database.
This is useful for apps which manage their own database and query
MediaProvider in order to synchronize between MediaProvider database and
their own database. When they don't have option to query all volumes,
they can misunderstand that rows of unmounted volume are deleted even
though the rows are not deleted in MediaProvider database.

Bug: 182734110
Test: com.android.providers.media.client.PublicVolumeTest#testIncludeRecentlyUnmountedVolume
Change-Id: I112c5e019a9c122f14fabe67aa1cbb793bd3a693