Return files-URI for inserts on hidden files
Hidden files are considered non-media files, but if callers try to
insert a hidden file using the uri
`content://media/external_primary/images/media`, then we return the
inserted row-id after appending to the passed URI. On the next scan, we
update the hidden file as non-media and the previously returned URI will
be invalid.

This change returns a URI which will be valid for the inserted non-media
file.

Bug:183065301
Test: atest android.scopedstorage.cts.host.LegacyStorageHostTest#testInsertHiddenFile

Signed-off-by: fengjinlan <fengjinlan@xiaomi.com>
Change-Id: I85eb7d7a60c42964435723406051f13be1a45910