Add User to LocalCallingIdentity.
As part of making the MediaProvider aware of multiple users, keep
tracking of the calling User in the LocalCallingIdentity. The User is
special in the sense that it is not cleared by
clearLocalCallingIdentity(), because even in places where we want to
clear the calling identity (eg scanning a volume), we need to be able to
retrieve the user later on to correctly resolve things like volume
names.

Bug: 182396009
Test: atest CtsScopedStorageDeviceTest
Test: atest MediaProviderTests

Change-Id: I98d5e90c2bb46e8085aeb793abbd187a5a6d7124