Avoid unnecessary FUSE node invalidation
The FUSE kernel driver is case-sensitive but the FUSE daemon needs to
provide a case-insensitive filesystem view. To workaround this
limitation, we need to ensure that there is only one dentry cache
entry representing any corresponding filename cases on the lower
filesystem. This requires us to do the following on a FUSE lookup with
mismatched cases:
1. Invalidate the previous node name with the
fuse_lowlevel_notify_inval_entry function
2. Invalidate the current node name by setting the entry_timeout to 0

In addition to doing the above, previous implementations also marked
the node as tainted so subsequent lookups will always do (2). This
was unnecessary since subsequent lookups with the nodes original
case would not need any additional invalidation while subsequent lookups
with a different case already triggered the invalidation logic in (1)
and (2)

This cl avoids marking the node as tainted and goes a step further to
update the original node name to the latest name received at
lookup. This significantly improves the performance in the following
scenario:
a. lookup foo once
b. lookup FOO many times

Without this change, the name in (b) will always differ from the
original name, in (a) hence we'd invalidate many times.

With the change, after the first lookup of FOO, we'd update the node
name to FOO so subsequent lookups of FOO will not require
invalidation. There is an edge case where we could trash and always
invalidate, e.g:

c. lookup foo once
d. lookup FOO once
e. lookup foo once
...

This is unlikely to happen and is not worse than the previous code
that marked nodes as tainted after the first invalidation

Test: atest
android.scopedstorage.cts.device.ScopedStorageDeviceTest#testCreateUpperCaseDeleteLowerCase
Bug: 183762925

Change-Id: I15a3b84535521cf61445856f5de5e12a1a1d4fbd