Add ItemsProvider#getCategories()
This API provides ViewModel with details about Categories in the initial
grid view for the photo picker.

getCategories() returns only basic information about the category like:
category name, number of items in the category and the uri of the most
recent item in the category.

The number of local categories for photopicker v0 are only 4:
Screenshots, Camera, Videos, Favorites. The API does not return the
category name if there are no items that belong to the category.

Bug: 185905178
Test: atest LocalItemsProviderTest
Change-Id: Ibc5c409c669bb646fd5d196ffa0fb3ae08f89d21