Allow volume paths from both users when verifying file columns.
assertFileColumnsConsistent() verifies that the value specified in the
DATA column matches up with the path of the volume. When there's
a clone profile, we don't always know whether a volume name like
external_primary refers to the volume of the owner user, or the clone
user. In particular, media scanning may happen in the context of the
owner user, yet it may be scanning a volume of a clone user.

To fix this, just return the paths for all users that we manage.

Bug: 183949139
Test: TEST_MAPPING
Change-Id: If999ac17c3492901a035092d1886c1ee123899c7