Add UserIdManager for photo picker to query UserIds
This class is essential for the caller of ItemsProvider to query and
specify UserId. This is also essential for the UI to know if Work
profile exists or not; which profile is the current profile etc.

The personal profile and managed (aka work) profile information is
cached and we listen to broadcasts for managed profile added/removed,
upon which we invalidate the cache.

The main client for this class would be the UI. (Please refer to the
class diagram in the design doc).

Bug: 188394433
Test: atest UserIdManagerTest
Change-Id: I4cbd3eed548961eff0cf3a223651312d5e5c236f