No-op refactor to support more than 2 dbs in DatabaseHelper
Currently, the DatabaseHelper assumes MediaProvider can only have 2
dbs, so if a db is NOT internal, it is external. For the photo picker,
we will introduce a separate db, picker.db which breaks this
assumption.

In preparation to introduce the picker.db, we refactor the code with
explicit methods to check if db is internal or external. In future cl,
we'll add an explicit check for picker db.

Test: presubmit
Bug: 190713331
Change-Id: Iab0fbe9a9047fc63a043685312a1e1fd394f9b48
Merged-In: Iab0fbe9a9047fc63a043685312a1e1fd394f9b48
(cherry picked from commit faa5427c1b511a0bf88192cf8e01d48321395524)